requirements.txt
requests
beautifulsoup4
PyPDF2
openai
python-dotenv
fastapi
uvicorn
reportlab
pytz